Address actor subsystem review feedback
