+++ ydb/library/actors/core/subsystems/cgroup/cgroup_v1.cpp
@@ -1,6 +1,7 @@
 #include "cgroup_v1.h"
 
 #include "cgroup_events.h"
+#include "cgroup_helpers.h"
 
 #include <ydb/library/actors/core/actor.h>
 #include <ydb/library/actors/core/actor_bootstrapped.h>
@@ -23,26 +24,10 @@ namespace NActors {
 
         constexpr ui64 V1UnlimitedThreshold = 1ULL << 60;
 
-        std::optional<TString> TryReadFile(const TFsPath& path) {
-            try {
-                return TFileInput(path.GetPath()).ReadAll();
-            } catch (...) {
-                return std::nullopt;
-            }
-        }
-
-        std::optional<ui64> TryReadUInt(const TFsPath& path) {
-            auto content = TryReadFile(path);
-            if (!content) {
-                return std::nullopt;
-            }
-
-            ui64 value = 0;
-            if (!TryFromString(StripString(*content), value)) {
-                return std::nullopt;
-            }
-            return value;
-        }
+        using NCGroupDetail::MakeCGroupDirectory;
+        using NCGroupDetail::ParseKeyValueFile;
+        using NCGroupDetail::TryReadFile;
+        using NCGroupDetail::TryReadUInt;
 
         std::optional<i64> TryReadInt(const TFsPath& path) {
             auto content = TryReadFile(path);
@@ -83,24 +68,6 @@ namespace NActors {
                 return std::nullopt;
             }
             return TCGroupV1Limit{.Value = parsed};
-        }
-
-        template<class TCallback>
-        void ParseKeyValueFile(const TString& content, TCallback&& callback) {
-            TStringInput input(content);
-            TString line;
-            while (input.ReadLine(line)) {
-                TVector<TString> fields;
-                StringSplitter(line).Split(' ').SkipEmpty().Collect(&fields);
-                if (fields.size() != 2) {
-                    continue;
-                }
-
-                ui64 value = 0;
-                if (TryFromString(fields[1], value)) {
-                    callback(TStringBuf(fields[0]), value);
-                }
-            }
         }
 
         struct TCGroupV1ControllerPaths {
@@ -162,13 +129,6 @@ namespace NActors {
             return result.Empty()
                 ? std::nullopt
                 : std::optional<TCGroupV1ControllerPaths>(std::move(result));
-        }
-
-        TFsPath MakeCGroupDirectory(const TString& root, const TString& cGroupPath) {
-            if (cGroupPath == "/") {
-                return TFsPath(root);
-            }
-            return TFsPath(root) / cGroupPath.substr(1);
         }
 
         struct TMemoryStatFields {
